Support interfaces in Records
- Fixes #265
