+++ api/src/main/java/org/jboss/forge/roaster/model/source/JavaRecordSource.java
@@ -9,7 +9,8 @@ public interface JavaRecordSource extends
          JavaSource<JavaRecordSource>,
          JavaRecord<JavaRecordSource>,
          MethodHolderSource<JavaRecordSource>,
-         TypeHolderSource<JavaRecordSource>
+         TypeHolderSource<JavaRecordSource>,
+         InterfaceCapableSource<JavaRecordSource>
 {
    JavaRecordComponentSource addRecordComponent(String type, String name);
    JavaRecordComponentSource addRecordComponent(Class<?> type, String name);
